Revert "lkl: Implement lkl_printf and lkl_bug internally"
This reverts commit 486afba4fe4e973df0c57635ba3969dd11e8699f.

lkl_printf is called before Linux is initialized in several instances
and this use-case is the main reason for this API's
existence. Unfortunately, and suprisingly, the Linux vsnprintf uses
WARN* macros which will prints information about the CPU, current
process, stack trace etc. and some of this information does require
Linux to be initialized.
